requirements.txt
flask>=3.0.0
pypdf>=4.0.0
openai>=1.0.0
sentence-transformers>=2.2.0
torch>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
playwright>=1.40.0
googlesearch-python>=1.2.0
